Added return statements to new migrations.

+++ data/migrations/20200323160349_components-addcustom.js
@@ -1,12 +1,19 @@
 
 exports.up = function(knex) {
-  knex.schema.table('components', tbl => {
+  return knex.schema.table('components', tbl => {
     tbl.string('custom')
   })
 };
 
 exports.down = function(knex) {
-  knex.schema.table('components', tbl => {
-      tbl.dropColumn('custom')
-  })
+  return knex.schema.table('components', tbl => {
+    knex.schema.hasColumn('components', 'custom')
+    .then(exists => {
+      if (exists) {
+        tbl.dropColumn('custom');
+      }
+    }).catch(err => {
+      console.log(err);
+    });
+  });
 };
