+++ data/migrations/20200323161417_jobsheets2.js
@@ -1,12 +1,19 @@
 
 exports.up = function(knex) {
-  knex.schema.table('jobsheets', tbl => {
-    tbl.boolean('completed').defautTo(false)
+  return knex.schema.table('jobsheets', tbl => {
+    tbl.boolean('completed').defaultTo(false)
   })
 };
 
 exports.down = function(knex) {
-  knex.schema.table('jobsheets', tbl => {
-    tbl.dropColumn('completed')
-  })
+  return knex.schema.table('jobsheets', tbl => {
+    knex.schema.hasColumn('jobsheets', 'completed')
+    .then(exists => {
+      if (exists) {
+        tbl.dropColumn('completed');
+      }
+    }).catch(err => {
+      console.log(err);
+    });
+  });
 };
